readded basic auth; moved auth tests to separate test class
