+++ samples/client/petstore/php/SwaggerClient-php/lib/Api/StoreApi.php
@@ -279,8 +279,6 @@ class StoreApi
         if ($apiKey !== null) {
             $headers['api_key'] = $apiKey;
         }
-        /**
-        */
         $query = \GuzzleHttp\Psr7\build_query($queryParams);
 
         try {
